requirements.txt
flask
pandas
scikit-learn
joblib
python-dotenv
google-generativeai